rename example_5 to polymorphism
