+++ polymorphism/include/polymorphism/consume_class_with_interface.hpp
@@ -1,7 +1,7 @@
 #ifndef CONSUME_CLASS_WITH_INTERFACE_HPP
 #define CONSUME_CLASS_WITH_INTERFACE_HPP
 
-#include <example_5/i_super_cool_features.hpp>
+#include <polymorphism/i_super_cool_features.hpp>
 
 namespace classic {
