+++ polymorphism/test/test_consume.cpp
@@ -1,9 +1,9 @@
 #include <ut.hpp>
 
-#include <example_5/consume_class_that_adheres_to_concept.hpp>
-#include <example_5/consume_class_with_interface.hpp>
-#include <example_5/impl_with_interface.hpp>
-#include <example_5/impl_without_interface.hpp>
+#include <polymorphism/consume_class_that_adheres_to_concept.hpp>
+#include <polymorphism/consume_class_with_interface.hpp>
+#include <polymorphism/impl_with_interface.hpp>
+#include <polymorphism/impl_without_interface.hpp>
 
 // ctest --rerun-failed --output-on-failure
